- Update object_store to 0.12.2
- Update build.rs to download from repo.hops.works/master/libhdfs
- Add -lresolv for macOS linking
- Fix GetRange::Suffix type (usize -> u64)
- Remove unused build dependencies (fs_extra, url)

+++ tests/test_hopsfs_object_store.rs
@@ -104,7 +104,7 @@ mod test {
             .await?;
 
         let get_opts = GetOptions {
-            range: Some(GetRange::Suffix(buf_le)),
+            range: Some(GetRange::Suffix(buf_le as u64)),
             ..Default::default()
         };
